Refactor request code (#21)
* Refactor request code to reduce code duplication

* Order methods and method names

Also use an union of constant strings instead of the unused type map.

* Refactor http request utils

* Improve type precision

Replace "any" type if possible.

* Add "tests" directory on lint command

* Add IReplaceableAccountId interface

"accountId" property now has jsdoc. It cannot be
null according to the jmap specifications, but the library will replace
it by a default value if it is.
Also refactor the replaceAccountId method so the object is not recreated
if no modification is done on it.